+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Enhancements
 
 - Add `repo` option to `Commanded.Projections.Ecto` macro ([#1](https://github.com/slashdotdash/commanded-ecto-projections/pull/1)).
+- Optional `after_update/3` callback function in projectors.